# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API da Livraria Libris",
    "description": "Documentação da API da livraria online Libris",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/livros/": {
      "get": {
        "tags": [
          "Livros"
        ],
        "summary": "Lista todos os livros",
        "description": "Retorna uma lista com todos os livros cadastrados.",
        "responses": {
          "200": {
            "description": "Lista de livros retornada com sucesso."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/livros/{id}": {
      "get": {
        "tags": [
          "Livros"
        ],
        "summary": "Retorna um livro específico",
        "description": "Retorna os detalhes de um livro com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livro retornado com sucesso."
          },
          "404": {
            "description": "Livro não encontrado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/livros/pesquisa/{termo}": {
      "get": {
        "tags": [
          "Livros - Pesquisa"
        ],
        "summary": "Retorna um livro específico por termo de pesquisa",
        "description": "Retorna os detalhes de um livro com base no termo de pesquisa fornecido.",
        "parameters": [
          {
            "name": "termo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de livros encontrada com sucesso."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login/": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Realiza o login de um cliente",
        "description": "Permite que um cliente faça login no sistema.",
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/login/": {
      "post": {
        "tags": [
          "Login Admin"
        ],
        "summary": "Realiza o login de um administrador",
        "description": "Permite que um administrador faça login no sistema.",
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/administrador/livros": {
      "post": {
        "tags": [
          "Livros Admin"
        ],
        "summary": "Cadastra um livro",
        "description": "Realiza o cadastro de um novo livro.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "O Senhor dos Anéis"
                },
                "isbn": {
                  "type": "string",
                  "example": "978-0-345-33968-9"
                },
                "preco": {
                  "type": "number",
                  "example": 29.99
                },
                "autor": {
                  "type": "string",
                  "example": "J.R.R. Tolkien"
                },
                "editora": {
                  "type": "string",
                  "example": "HarperCollins"
                },
                "categoria": {
                  "type": "string",
                  "example": "Fantasia"
                },
                "capa": {
                  "type": "string",
                  "example": "https://example.com/capa.jpg"
                },
                "descricao": {
                  "type": "string",
                  "example": "Um clássico da literatura fantástica."
                },
                "anoPublicacao": {
                  "type": "number",
                  "example": 1954
                },
                "estoque": {
                  "type": "number",
                  "example": 10
                },
                "adminId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Livro cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/administrador/livros/{id}": {
      "delete": {
        "tags": [
          "Livros Admin"
        ],
        "summary": "Exclui um livro",
        "description": "Realiza a exclusão de um livro existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livro excluído com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Livros Admin"
        ],
        "summary": "Atualiza um livro",
        "description": "Realiza a atualização de um livro existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "O Senhor dos Anéis"
                },
                "isbn": {
                  "type": "string",
                  "example": "978-0-345-33968-9"
                },
                "preco": {
                  "type": "number",
                  "example": 29.99
                },
                "autor": {
                  "type": "string",
                  "example": "J.R.R. Tolkien"
                },
                "editora": {
                  "type": "string",
                  "example": "HarperCollins"
                },
                "categoria": {
                  "type": "string",
                  "example": "Fantasia"
                },
                "capa": {
                  "type": "string",
                  "example": "https://example.com/capa.jpg"
                },
                "descricao": {
                  "type": "string",
                  "example": "Um clássico da literatura fantástica."
                },
                "anoPublicacao": {
                  "type": "number",
                  "example": 1954
                },
                "estoque": {
                  "type": "number",
                  "example": 10
                },
                "adminId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Livro atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}